+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.9.2 - TBD
  * Added `API::emptyFolder(Type\FolderIdType $folderId, $deleteType = 'SoftDelete', $deleteSubFolders = false, array $options = [])`
  * Added `MailAPI::emptyTrash(array $options = [])`
+ * Added `CalendarAPI::getAvailabilityFor($startTime, $endTime, array $users, array $options = [])`
+ * Added `CalendarAPI::areAvailable($startTime, $endTime, $period, array $users, array $options = [])`
  
 ## 0.9.1 - 2016-11-28
  * Changed the default ItemShape of `listChanges` to `AllProperties`